Tests: fix mockDAGSyncer
I had observed that increasing the number of items or replicas in the tests
resulted in an exponential time walking the DAGs. It turns out that fetched
blocks were not marked for "HasBlock" and always returned false, which meant
that the full dag was always being re-walked on every update. High branching
factor due to concurrent updates to the DAG meant re-walking many times on
every update and never pruning branches except when hitting a head.

This had not effect on results, it just made the tests suck.