Handle one received block at at time.
Receiving a new block on top of a previosly received blocked that is still
being processed may result in the former root staying as head even if it
should have been replaced by the new block (which is an additional root).

This means that we produce spurious heads in a non-controlled fashion, which
may result in overhead when everyone is committing since the number of heads
in every peer will depend on how fast new blocks are processed.

This should not affect the contents of the store, which do converge, but just
makes the DAG differ more than needed thus increasing need for sync.

The solution here is to track which root-nodes are being processed and, check
if another branch for other root nodes arrives to them. In that case, it
should not continue processing until the previous one is done. Once that happens
it should be able to detect the previous root node as a head and replace it
accordingly rather than resulting in two heads.