Fix: batch value and priority setting correctly
We were batching the setting of the elems and tombstones keys, but writing
values and priorities associated to them directly to the database.

Thus, whenever we List() or Get() elements, they would have had a value, but
where not in the Elems set (which was still batching). InSet() returned false,
and the Element() method cleared the value.