Improve the re-broadcasting mechanism
Instead of skipping re-broadcasting whenever some new head has been
received in the interval, this will remember the CIDs received
during the re-broadcast interval and proceed to broadcast only
those that have not been seen during that interval.

This avoids a situation where a peer with new heads does
not rebroadcast them because the peers with old heads are rebroadcasting
the old ones.