Broadcast a protobuf message, not a CID.
This switches to broadcasting a protobuf message rather than a CID while
keeping backwards compatibility with Cid-bytes broadcasts.

Broadcasting a protobuf will allows us to expand the message in the future and
re-broadcast messages that that includes signatures from trusted peers, so
that other peers can get meaninful updates when joining a cluster even when
trusted peers are offline. For this the introduction of message validators
directly in go-ds-crdt will be necessary.