Disable multi-head processing by default and add option
The reason why it was disabled in the previous version is that it may cause
excessive branching. For example:

Processing head A should make A one of the current head.

If a new head B that comes on top of A arrives, processing it may finish
before A became a head itself. This will make B a head on its own, and A will
also be a head later (which should have been replaced).

Thus we end up with more heads than needed, which results in branches when we
issue a new head.