Set Heads immediately when no heads are known
If a new node does not have any heads and receives some, they are set as heads
immediately, and processed as before. This means that any update made by that
node will be built on top of those seen heads, without the node having to have
fully synced (the moment heads are created).

Currently the node would be syncing and upon seeing new heads, create orphan
branches with priority starting at 0. This means an update from a new node
would never have enough priority to overwrite older updates to the same key
from other nodes while the new node is not fully synced, which is not expected
behaviour.