Use "values" and a "priorities" namespaces instead of key suffixes
Upon review on how datastore values and priorities are stored, I could not
find any sense in the previous approach. It makes more sense, I think, to have
a namespace to store values associated to each key, and a namespace to store
priorities.

This makes listing Elements faster as it reduces the amount of Keys that need
to be traversed by 50%.